+++ packages/web/app/features/Invoice/InvoiceForm/InvoiceItemCells/CodeCell.jsx
@@ -1,21 +1,9 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Box } from '@material-ui/core';
-
-const StyledItemCell = styled(Box)`
-  .MuiFormHelperText-root {
-    font-size: 14px;
-  }
-`;
-
-const ViewOnlyCell = styled.div`
-  display: flex;
-  font-size: 14px;
-  padding-left: 15px;
-`;
+import { ViewOnlyCell } from './ViewOnlyCell';
+import { ItemCell } from './ItemCell';
 
 export const CodeCell = ({ item }) => (
-  <StyledItemCell width="10%">
+  <ItemCell width="10%">
     <ViewOnlyCell>{item.productCode}</ViewOnlyCell>
-  </StyledItemCell>
+  </ItemCell>
 );
