+++ packages/web/app/features/Invoice/InvoiceForm/InvoiceItemCells/ItemCell.jsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { Box } from '@mui/material';
+
+export const ItemCell = styled(Box)`
+  .MuiFormHelperText-root {
+    font-size: 14px;
+  }
+`;
