+++ packages/web/app/features/Invoice/InvoiceForm/InvoiceItemCells/OrderedByCell.jsx
@@ -1,23 +1,7 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Box } from '@material-ui/core';
-import {
-  AutocompleteField,
-  Field,
-  NoteModalActionBlocker,
-} from '../../../../components';
-
-const StyledItemCell = styled(Box)`
-  .MuiFormHelperText-root {
-    font-size: 14px;
-  }
-`;
-
-const ViewOnlyCell = styled.div`
-  display: flex;
-  font-size: 14px;
-  padding-left: 15px;
-`;
+import { AutocompleteField, Field, NoteModalActionBlocker } from '../../../../components';
+import { ViewOnlyCell } from './ViewOnlyCell';
+import { ItemCell } from './ItemCell';
 
 export const OrderedByCell = ({
   index,
@@ -26,7 +10,7 @@ export const OrderedByCell = ({
   practitionerSuggester,
   handleChangeOrderedBy,
 }) => (
-  <StyledItemCell width="19%" data-testid="styleditemcell-tfvb">
+  <ItemCell width="19%" data-testid="styleditemcell-tfvb">
     {isItemEditable ? (
       <NoteModalActionBlocker>
         <Field
@@ -42,5 +26,5 @@ export const OrderedByCell = ({
     ) : (
       <ViewOnlyCell>{item?.orderedByUser?.displayName}</ViewOnlyCell>
     )}
-  </StyledItemCell>
+  </ItemCell>
 );
