+++ packages/web/app/features/Invoice/InvoiceForm/InvoiceItemCells/QuantityCell.jsx
@@ -1,26 +1,10 @@
 import React from 'react';
-import styled from 'styled-components';
-import { Box } from '@material-ui/core';
-import {
-  Field,
-  NumberField,
-  NoteModalActionBlocker,
-} from '../../../../components';
-
-const StyledItemCell = styled(Box)`
-  .MuiFormHelperText-root {
-    font-size: 14px;
-  }
-`;
-
-const ViewOnlyCell = styled.div`
-  display: flex;
-  font-size: 14px;
-  padding-left: 15px;
-`;
+import { Field, NumberField, NoteModalActionBlocker } from '../../../../components';
+import { ViewOnlyCell } from './ViewOnlyCell';
+import { ItemCell } from './ItemCell';
 
 export const QuantityCell = ({ index, item, isItemEditable }) => (
-  <StyledItemCell width="10%" paddingLeft="24px">
+  <ItemCell width="10%" paddingLeft="24px">
     {isItemEditable ? (
       <NoteModalActionBlocker>
         <Field
@@ -41,5 +25,5 @@ export const QuantityCell = ({ index, item, isItemEditable }) => (
     ) : (
       <ViewOnlyCell>{item?.quantity}</ViewOnlyCell>
     )}
-  </StyledItemCell>
+  </ItemCell>
 );
