+++ packages/web/app/features/Invoice/InvoiceForm/InvoiceItemCells/ViewOnlyCell.jsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const ViewOnlyCell = styled.div`
+  display: flex;
+  font-size: 14px;
+  padding-left: 15px;
+`;
